Add InitModels, TrainModel and CalcObjFunc.m

+++ GeneratePattern.m
@@ -1,14 +1,20 @@
-function [ output_args ] = GeneratePattern( patternSize )
+function [ returnStatus ] = GeneratePattern( patternSize )
 %GeneratePattern Generate default patterns and save them to .mat files.
+%   If the pattern files exist, then skip this function.
 
-%%Pattern Cross
+%Check pattern files.
+if exist('pattern1.mat', 'file') ~= 0 && exist('pattern2.mat', 'file') ~= 0
+    return;
+end
+
+%Pattern Cross
 pattern = ones(patternSize);
 pattern(30:35, 17:48) = 0;
 pattern(17:48, 30:35) = 0;
 imwrite(pattern, 'pattern1.bmp');
 save('pattern1.mat', 'pattern');
 
-%%Pattern Triangle
+%Pattern Triangle
 pattern = ones(patternSize);
 marginX = round(patternSize / 2);
 marginY = 21;
@@ -23,6 +29,8 @@ for y = marginY:patternSize-marginY
 end
 imwrite(pattern, 'pattern2.bmp');
 save('pattern2.mat', 'pattern');
+
+returnStatus = true;
 
 end
